Use composition
