+++ flink-runtime/src/main/java/org/apache/flink/runtime/util/BashJavaUtils.java
@@ -23,6 +23,7 @@ import org.apache.flink.configuration.TaskManagerOptions;
 import org.apache.flink.runtime.clusterframework.TaskExecutorProcessSpec;
 import org.apache.flink.runtime.clusterframework.TaskExecutorProcessUtils;
 import org.apache.flink.runtime.taskexecutor.TaskManagerRunner;
+import org.apache.flink.runtime.util.config.memory.ProcessMemoryUtils;
 
 import java.util.Arrays;
