+++ flink-runtime/src/main/java/org/apache/flink/runtime/util/config/memory/ProcessMemory.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.runtime.util.config.memory;
+
+/**
+ * Process memory components.
+ */
+public class ProcessMemory<IM extends FlinkMemory> {
+	private final IM flinkMemory;
+	private final JvmMetaspaceAndOverhead jvmMetaspaceAndOverhead;
+
+	ProcessMemory(IM flinkMemory, JvmMetaspaceAndOverhead jvmMetaspaceAndOverhead) {
+		this.flinkMemory = flinkMemory;
+		this.jvmMetaspaceAndOverhead = jvmMetaspaceAndOverhead;
+	}
+
+	public IM getFlinkMemory() {
+		return flinkMemory;
+	}
+
+	public JvmMetaspaceAndOverhead getJvmMetaspaceAndOverhead() {
+		return jvmMetaspaceAndOverhead;
+	}
+}
